==> 02_willow.qmd <==
---
title: "02_willow"
format: html
editor: visual
---

## 02_Willow

Time and distance removal model

```{r clear space}
# Clean your workspace to reset your R environment
rm( list = ls() )
# Check that you are in the right project folder
getwd()
```

## Library

```{r library}

library(tidyr)
library(dplyr)
library(unmarked)
library(raster)
library(rgdal)
library(ggplot2)

```

## Load data

```{r load data}

# WIPT male count data:
all.obs <- read.csv("E:/gyrf_analysis/MSgyrfalcon/MSgyrfalcon/all_obs_17Oct23.csv")
all.obs <- all.obs[,-c(1)]

all.obs$exact_distance[all.obs$species != "WIPT"] <- NA
  

wipt.only <- filter(all.obs, species == "WIPT")

# adjust these labels so sex matches number observed
wipt.only[247, 10] = "MMMF"
wipt.only[249, 10] = "MMF"

#remove single F and single U

males <- filter(wipt.only, sex != "F")
males2 <- filter(males, sex != "U")


males2$sex[males2$sex == "M,F"] <- "MF"
males2$sex[males2$sex == "F,M"] <- "MF"
males2$sex[males2$sex == "M,2U"] <- "MUU"


# remove females from grouped counts
males2$count <- as.numeric(males2$count)
males2$sex <- as.factor(males2$sex)

#MF
males2$count[males2$sex == "MF"] <- males2$count[males2$sex == "MF"] - 1

#MMF
males2$count[males2$sex == "MMF"] <- males2$count[males2$sex == "MMF"] - 1

#MMMF
males2$count[males2$sex == "MMMF"] <- males2$count[males2$sex == "MMMF"] - 1

#MUU
males2$count[males2$sex == "MUU"] <- males2$count[males2$sex == "MUU"] - 2


# now take only columns needed then uncount groups

males3 <- males2[,c(1,4,5,8)]

males4 <- males3 %>%
   uncount(count)

# fix values that were NA but distances were given in the distance_category column
males4[113,3] <- 200
males4[218,3] <- 400
males4[220,3] <- 400
males4[222,3] <- 200


males5 <- data.frame(males4)
write.csv(males4, "E:/gyrf_analysis/MSgyrfalcon/MSgyrfalcon/wiptmalesonly_timeanddist_19Oct23.csv")



##############################

env.s <- read.csv("E:/gyrf_analysis/MSgyrfalcon/MSgyrfalcon/wipt_male_df_timeanddist_18Oct23.csv")

```

## Create unmarkedFrameGDR

### Distance Matrix

```{r distance matrix}

# create a separate df to manipulate for distance data
males6 <- data.frame(males5)

# now i need to get exact distances to 'binned data'


my.bins <- seq(0, 400, by = 50)

males6$distance_bin <- cut(males6$exact_distance, breaks = my.bins, include.lowest = TRUE)

# add empty sites (sites that had no WIPT male observations)

# first figure out what sites are missing
obs.ids <- as.data.frame(males6$id)
all.ids <- as.data.frame(env.s$id)

missing_df <- as.data.frame(setdiff(all.ids$`env.s$id`, obs.ids$`males6$id`))
colnames(missing_df)[1] <- "id" # change the name of the first column to 'id'

# now need to create dataframe that attaches empty sites to end of sites where ptarmigan were observed so i can then match those 'levels' of sites to the data frame of only observed. It make sure the formatting code knows there are sites where ptarmigan weren't observed. But the empty sites need to be attached to the end of the observed sites in an actual dataframe and then those levels matched to the observed dataframe as 'ghost' levels or else the order will be messed up and it will put the 'observed' ptarmigan into sites that were actually empty.

# so say the observed dataframe is df.obs = 'a, b, a, c, c, d' and you also had a site 'e' that was empty that you want to add to the levels. so a dataframe with all sites/levels, df.allsites = 'a, b, c, d, e'

# if you were to try and match the levels like 'levels(df.obs$id) <- levels(df.allsites$id) it would change the order of the observed sites to 'a, b, c, d, e' so then e would be matched up with a row that has an observation. you need to create a df that matches the same pattern of repeated site observations then add all the missing ones at the end: df.match = 'a, b, a, c, c, d, e' so then when you match them up the first sequence of 'a, b, a, c, c, d' matches up, then e is put at the end with empty values


# create my empty data frame I'm going to attach

num_columns <- 4 # define number of columns
num_rows <- 738 #define number of rows(the number of empty sites I'm adding)
col_names <- colnames(males6) # define what column names will be (the same as the frame i'm binding too)

# create the dataframe wtih my specifications
my_df <- data.frame(matrix(0, nrow = num_rows, ncol = num_columns))

# and add column names
colnames(my_df) <- col_names

# and replace 'id' column with names of all empty sites
my_df$id <- missing_df$id

# need to 'add' missing factor levels to the dataframe I'll actually be putting into the formatting code (creates 'ghost' sites for all empty sites)


males6$id <- as.factor(males6$id)
my_df$id <- as.factor(my_df$id)

levels(males6$id) <- c(levels(males6$id), levels(my_df))
levels(males6$id) # should be 988

########### format distance data for frame########
#make my bin breaks into a list so i can just put the object into the formatting insteading having to list them all out
my.bin.list <- as.list(my.bins)

# format distance data for the unmarkedFrameGDR:
yDist <- formatDistData(males6, distCol = "exact_distance", transectNameCol = "id", dist.breaks = my.bin.list)


sum(yDist)

```

### Time Removal Matrix

```{r time try 2}

# can use same dataframe but just specify its for the time-intervals

yRem <- formatDistData(males6, distCol = "time_int", transectNameCol = "id", dist.breaks = c(0, 2, 4, 6, 8, 10))

sum(yRem) # make sure its 485
```

### full Frame

```{r unmarkedFrameGDR}

# not sure if this makes a difference but put env covs in same order as yDist and yRem

ydist.df <- data.frame(yDist)
# Extract the row names of the matrix
row_names <- rownames(ydist.df)

# Add the row names as a new column in the dataframe
ydist.df$id <- row_names

# make sure both are factors
ydist.df$id <- as.factor(ydist.df$id)
env.s$id <- as.factor(env.s$id)

??left_join
covs <- full_join(ydist.df, env.s, by = 'id')

# add area to covs data
# area of the survey area was a 400m radius circle
# standardize the value by dividing by 10000

covs$area <- pi*400*400/10000


# site covariates
siCovs <- covs[,c(9, 16:30, 44)]
siCovs$id <- as.factor(siCovs$id)

# in the frame these are actually the yearlySiteCovs
oCovs <- covs[,c(9, 31:43)]
oCovs$id <- as.factor(oCovs$id)
oCovs$observer <- as.factor(oCovs$observer)

breaks <- seq(0, 400, by = 50)





umfGDR <- unmarkedFrameGDR(yDistance = yDist, yRemoval = yRem, numPrimary = 1, siteCovs = siCovs, obsCovs = NULL, yearlySiteCovs = oCovs, dist.breaks = breaks, unitsIn = 'm', period.lengths = NULL)


saveRDS(umfGDR, "unmarkedFrameGDR_wiptmale_20oct23.rds")



```

## Distance Removal Model: Zero Inflated Poisson (ZIP) vs Poisson

using function gdistremoval() in 'unmarked'

```{r gdr zip model}

# lambda:abundance, phi:availability (X - don't have these), removal:removal probablity covariates (these are my observer, date, etc), distance: detection function covariates

# data: object of class unmarkedFrameGDR
# keyfun: detection functions : "halfnorm", "hazard", "exp", or "uniform"
# output: "abund" or "density"
# unitsOut: units of density. either "ha" or "kmsq"
# mixture: either "P" "NB" or "ZIP"
# K: interger value speicying upper bound used in the integration
# starts: vector of starting values for model parameters
# methdod: optimization method used by optim
# se: logical specifying whether or not to comput standard errors
# engine:either "C" to use C++ code or "TMB" to use TMB for optimization
# threads: more stuff about C++... OpenMP

# set then standardize (dividing by 10,000) the area that was surveyed


# my abundance covariates for wipt: tundra, tussock, tall shrub, elevation

#### Output of density

umzip <- gdistremoval(lambdaformula = ~1 + tundra + tussock + tallshrub + elev + offset(log(area)), phiformula = ~1, removalformula = ~1 + julian + min_after_sun + wind + observer, data = umfGDR, keyfun = "halfnorm", output = "density", unitsOut = "kmsq", mixture = "ZIP")


summary(umzip)


ump <- gdistremoval(lambdaformula = ~1 + tundra + tussock + tallshrub + elev + offset(log(area)), phiformula = ~1, removalformula = ~1 + julian + min_after_sun + wind + observer, data = umfGDR, keyfun = "halfnorm", output = "density", unitsOut = "kmsq", mixture = "P")

summary(ump)

### Output of abundance, but with the addition of area offset it's abundance

zip.abund <- gdistremoval(lambdaformula = ~1 + tundra + tussock + tallshrub + elev + offset(log(area)), phiformula = ~1, removalformula = ~1 + julian + min_after_sun + wind + observer, data = umfGDR, keyfun = "halfnorm", output = "abund", mixture = "ZIP")
summary(zip.abund)

p.abund <- gdistremoval(lambdaformula = ~1 + tundra + tussock + tallshrub + elev + offset(log(area)), phiformula = ~1, removalformula = ~1 + julian + min_after_sun + wind + observer, data = umfGDR, keyfun = "halfnorm", output = "abund", mixture = "P")
summary(p.abund)
summary(zip.abund)
```

## Spatial predictions

```{r}

# First need to scale study-wide rasters to match scaling of model:


elev <- raster("E:/gyrf_analysis/MSgyrfalcon/dem_1k.tif")

tallshrub <- raster("E:/gyrf_analysis/MSgyrfalcon/MSgyrfalcon/1k prediction rasters/tallshrub.tif")

tundra <- raster("E:/gyrf_analysis/MSgyrfalcon/MSgyrfalcon/1k prediction rasters/tundra.tif")

tussock <- raster("E:/gyrf_analysis/MSgyrfalcon/MSgyrfalcon/1k prediction rasters/tussock.tif")



# standardize by each rasters mean and standard deviation

# elev
cellStats(elev, mean) #173.7808
cellStats(elev, "sd") #120.5669

elev.s <- (elev-173.7808)/120.5669

# tundra
cellStats(tundra, mean) #0.1608737
cellStats(tundra, sd) #0.08171017

tundra.s <- (tundra-0.1608737)/0.08171017

# tussock
cellStats(tussock, mean) #0.1202638
cellStats(tussock, sd) #0.08615176

tussock.s <- (tussock-0.1202638)/0.08615176

# tallshrub
cellStats(tallshrub, mean)
cellStats(tallshrub, sd)

tallshrub.s <- (tallshrub - 0.1321896)/0.07754615


area.raster <- elev 
values(area.raster) <- 1000*1000/10000 # area of a grid pixel, divided by 10000 to standardize

# pull scaled observation/yearlySite covs from dataframe
jul.raster <- elev # create a raster using one thats already in correct crs/scale
values(jul.raster) <- mean(env.s$julian) # replace data with that of what you want - doing it with the mean scaled value of obs covs

wind.raster <- elev 
values(wind.raster) <- mean(env.s$wind)

minsun.raster <- elev 
values(minsun.raster) <- mean(env.s$min_after_sun)

obs.raster <- elev 
values(obs.raster) <- as.factor("DS")



pred.surface <- stack(tundra.s, tussock.s, tallshrub.s, elev.s, area.raster, jul.raster, wind.raster, minsun.raster, obs.raster)

names(pred.surface) <- c("tundra", "tussock", "tallshrub", "elev", "area", "julian", "wind", "min_after_sun", "observer")


# now predict to study area!
ump.prediction <- predict(ump, type="lambda", newdata=pred.surface)
plot(ump.prediction)
cellStats(ump.prediction, "sum") # 50mil



umzip.prediction <- predict(umzip, type = "lambda", newdata = pred.surface)
plot(umzip.prediction)
cellStats(umzip.prediction, "sum") # 50 mil


pabund.prediction <- predict(p.abund, type = "lambda", newdata = pred.surface)
plot(pabund.prediction)
cellStats(pabund.prediction, "sum") # 26 059, 196 mil
writeRaster

zipabund.prediction <- predict(zip.abund, type = "lambda", newdata = pred.surface)
plot(zipabund.prediction)
cellStats(zipabund.prediction, "sum") # 26 923, 906 mil



# Write the raster to a GeoTIFF file
writeRaster(zipabund.prediction, filename = names(zipabund.prediction), bylayer = TRUE, format = "GTiff", overwrite = TRUE)


writeRaster(pabund.prediction, filename = names(pabund.prediction), bylayer = TRUE, format = "GTiff", overwrite = TRUE)
```

## Abundance Model Partial Prediction Plots

```{r prep ppp}
# Estimate partial prediction plots for predictors with 95% CIs not overlapping zero:

# Start by creating our datasets to predict over

# how many values do we use:
n <- 100

# Use the observed values to define our range:

# import unscaled predictors
env <- read.csv("E:/gyrf_analysis/MSgyrfalcon/MSgyrfalcon/final_env_join_18Oct23.csv")

elevation.pp <- seq( min( env[,"elev"]),max(500),
                   length.out = n )

tallshrub.pp <- seq( min( env[,"tallshrub"]),max( env[,"tallshrub"]),
                   length.out = n )

tundra.pp <- seq( min( env[,"tundra"]),max( env[,"tundra"]),
                   length.out = n )

tussock.pp <- seq( min( env[,"tussock"]),max( env[,"tussock"]),
                   length.out = n )

#standardize predictors:
elev.std <- scale( elevation.pp )
tallshrub.std <- scale( tallshrub.pp )
tundra.std <- scale( tundra.pp )
tussock.std <- scale( tussock.pp )
area.x = 1000*1000/10000
```

### Elevation

```{r elevation ppp}
### ELEVATION

#combine standardized predictors into a new dataframe to predict partial relationship
# for abundance submodel:

elevData <- data.frame( tundra = 0, tussock = 0, tallshrub = 0, elev = elev.std, area = area.x )

#predict partial relationship:
pred.elev.p <- predict(p.abund, type = "lambda", newdata = elevData, 
                          appendData = TRUE )

pred.elev.zip <- predict(zip.abund, type = "lambda", newdata = elevData, 
                          appendData = TRUE )

#view
head( pred.elev.p ); dim( pred.elev.p )
head( pred.elev.zip); dim( pred.elev.zip)

### plot

elev.p <- cbind( pred.elev.p[,c("Predicted", "lower", "upper") ], elevation.pp ) %>%
  # define x and y values
  ggplot(., aes( x = elevation.pp, y = Predicted ) ) + 
  #choose preset look
  theme_classic( base_size = 30) +
  # add labels
 # ylim(0, 13) +
  labs( x = "Elevation (m)", y = "Willow Ptarmigan per kmsq" ) +
  theme(text=element_text(size=65)) +
    # add band of confidence intervals
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )


#view
elev.p
ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_elev_poiss.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)


# ZIP ppp
elev.zip <- cbind( pred.elev.zip[,c("Predicted", "lower", "upper") ], elevation.pp ) %>%
  # define x and y values
  ggplot(., aes( x = elevation.pp, y = Predicted ) ) + 
  #choose preset look
  theme_classic( base_size = 35) +
  # add labels
 # ylim(0, 13) +
  # add labels
  labs( x = "Elevation (m)", y = "Willow Ptarmigan per kmsq" ) +
  theme(text=element_text(size=65)) +
  # add band of confidence intervals
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )


#view
elev.zip
ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_elev_zip.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)

```

###  Tundra

```{r tundra ppp}

### TUNDRA

#combine standardized predictors into a new dataframe to predict partial relationship
# for abundance submodel:
tundData <- data.frame( tundra = tundra.std, tussock = 0, tallshrub = 0, elev = 0, area = area.x )

#predict partial relationship:
tun.p <- predict( p.abund, type = "lambda", newdata = tundData, 
                          appendData = TRUE )
tun.zip <- predict( zip.abund, type = "lambda", newdata = tundData, 
                          appendData = TRUE )

#view
head( tun.p ); dim( tun.p )
head( tun.zip ); dim( tun.zip)


### plot

tund.p <- cbind( tun.p[,c("Predicted", "lower", "upper") ], tundra.pp ) %>%
  # define x and y values
  ggplot(., aes( x = tundra.pp, y = Predicted ) ) + 
  #choose preset look
  #ylim(0, 45) +
  theme_classic( base_size = 30) +
  # turn x axis into percent scale
  scale_x_continuous(labels = scales::percent_format(accuracy=1)) +
  # add labels
  labs( x = "Tundra", y = "Willow Ptarmigan per kmsq" ) +
  # add band of confidence intervals
  theme(text=element_text(size=65)) +
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue")


#view
tund.p
ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_tundra_poiss.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)

### ZIP

tund.zip <- cbind( tun.zip[,c("Predicted", "lower", "upper") ], tundra.pp ) %>%
  # define x and y values
  ggplot(., aes( x = tundra.pp, y = Predicted ) ) + 
  #choose preset look
  #ylim(0, 45) +
  theme_classic( base_size = 30) +
  # turn x axis into percent scale
  scale_x_continuous(labels = scales::percent_format(accuracy=1)) +
  # add labels
  labs( x = "Tundra", y = "Willow Ptarmigan per kmsq" ) +
  # add band of confidence intervals
  theme(text=element_text(size=65)) +
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue")


#view
tund.zip
ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_tundra_zip.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)

```

### Tussock

```{r tussock ppp}
### TUSSOCK

#combine standardized predictors into a new dataframe to predict partial relationship
# for abundance submodel:
tussData <- data.frame( tundra = 0, tussock = tussock.std, tallshrub = 0, elev = 0, area = area.x )

#predict partial relationship:
tuss.p <- predict( p.abund, type = "lambda", newdata = tussData, 
                          appendData = TRUE )

tuss.zip <- predict( zip.abund, type = "lambda", newdata = tussData, 
                          appendData = TRUE )
#view
head( tuss.p ); dim( tuss.p )
head( tuss.zip ); dim( tuss.zip )


### plot

p.tuss <- cbind( tuss.p[,c("Predicted", "lower", "upper") ], tussock.pp ) %>%
  # define x and y values
  ggplot(., aes( x = tussock.pp, y = Predicted ) ) + 
  #choose preset look
  theme_classic( base_size = 30) +
  # add labels
  #ylim(0, 45) +
  # turn x axis into percent scale
  scale_x_continuous(labels = scales::percent_format(accuracy=1)) +
  labs( x = "Tussock", y = "Willow Ptarmigan per kmsq" ) +
  # add band of confidence intervals
  theme(text=element_text(size=65)) +
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )
#view
p.tuss
ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_tussock_poiss.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)

### ZIP

zip.tuss <- cbind( tuss.zip[,c("Predicted", "lower", "upper") ], tussock.pp ) %>%
  # define x and y values
  ggplot(., aes( x = tussock.pp, y = Predicted ) ) + 
  #choose preset look
  theme_classic( base_size = 30) +
  # add labels
  #ylim(0, 45) +
  # turn x axis into percent scale
  scale_x_continuous(labels = scales::percent_format(accuracy=1)) +
  labs( x = "Tussock", y = "Willow Ptarmigan per kmsq" ) +
  # add band of confidence intervals
  theme(text=element_text(size=65)) +
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )
#view
zip.tuss

ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_tussock_zip.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)

```

### Tallshrub

```{r}
### ALL SHRUB

#combine standardized predictors into a new dataframe to predict partial relationship
# for abundance submodel:
tallData <- data.frame( tundra = 0, tussock = 0, tallshrub = tallshrub.std, elev = 0, area = area.x )

#predict partial relationship:
tall.p <- predict( p.abund, type = "lambda", newdata = tallData, 
                          appendData = TRUE )

tall.zip <- predict( zip.abund, type = "lambda", newdata = tallData, 
                          appendData = TRUE )
#view
head( tall.p ); dim( tall.p )
head( tall.zip ); dim( tall.zip )

### plot

p.tall <- cbind( tall.p[,c("Predicted", "lower", "upper") ], tallshrub.pp ) %>%
  # define x and y values
  ggplot(., aes( x = tallshrub.pp, y = Predicted ) ) + 
  #choose preset look
  #ylim(0, 45) +
  theme_classic( base_size = 30 ) +
  # turn x axis into percent scale
  scale_x_continuous(labels = scales::percent_format(accuracy=1)) +
  # add labels
  labs( x = "Tall shrub", y = "Willow Ptarmigan per kmsq" ) +
  # add band of confidence intervals
  theme(text=element_text(size= 65)) +
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )
#view
p.tall

ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_tallshrub_poiss.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)


### ZIP

zip.tall <- cbind( tall.zip[,c("Predicted", "lower", "upper") ], tallshrub.pp ) %>%
  # define x and y values
  ggplot(., aes( x = tallshrub.pp, y = Predicted ) ) + 
  #choose preset look
  #ylim(0, 45) +
  theme_classic( base_size = 30 ) +
  # turn x axis into percent scale
  scale_x_continuous(labels = scales::percent_format(accuracy=1)) +
  # add labels
  labs( x = "Tall shrub", y = "Willow Ptarmigan per kmsq" ) +
  # add band of confidence intervals
  theme(text=element_text(size= 65)) +
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )
#view
zip.tall

ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_tallshrub_zip.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)

```

## Density Model Partial Prediction Plots

### Elevation

```{r density model ppp}
### ELEVATION

#combine standardized predictors into a new dataframe to predict partial relationship
# for abundance submodel:

elevData <- data.frame( tundra = 0, tussock = 0, tallshrub = 0, elev = elev.std, area = area.x )

#predict partial relationship:
pred.elev.pd <- predict(ump, type = "lambda", newdata = elevData, 
                          appendData = TRUE )

pred.elev.zipd <- predict(umzip, type = "lambda", newdata = elevData, 
                          appendData = TRUE )

#view
head( pred.elev.pd ); dim( pred.elev.pd )
head( pred.elev.zipd); dim( pred.elev.zipd)

### plot

elev.pd <- cbind( pred.elev.pd[,c("Predicted", "lower", "upper") ], elevation.pp ) %>%
  # define x and y values
  ggplot(., aes( x = elevation.pp, y = Predicted ) ) + 
  #choose preset look
  theme_classic( base_size = 30) +
  # add labels
 # ylim(0, 13) +
  labs( x = "Elevation (m)", y = "Willow Ptarmigan density (kmsq" ) +
  theme(text=element_text(size=65)) +
    # add band of confidence intervals
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )


#view
elev.pd
ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_elev_poiss_density.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)


# ZIP ppp
elev.zipd <- cbind( pred.elev.zipd[,c("Predicted", "lower", "upper") ], elevation.pp ) %>%
  # define x and y values
  ggplot(., aes( x = elevation.pp, y = Predicted ) ) + 
  #choose preset look
  theme_classic( base_size = 35) +
  # add labels
 # ylim(0, 13) +
  # add labels
  labs( x = "Elevation (m)", y = "Willow Ptarmigan density (kmsq)" ) +
  theme(text=element_text(size=65)) +
  # add band of confidence intervals
  geom_smooth( aes(ymin = lower, ymax = upper ), 
               stat = "identity",
               size = 5, alpha = 0.5, fill = "lightblue3" ) +
  # add mean line on top
  geom_line( size = 5, color = "darkblue" )


#view
elev.zipd
ggsave("E:/gyrf_analysis/partial prediction plots 20oct23/wp_elev_zip_density.png", width = 20, height = 15, dpi = 300, limitsize = FALSE)
```

### Tundra

### Tussock

### Tallshrub
