==> 04_createpointsshapefile.Rmd <==
---
title: "04_createpointsshapefile"
author: "Michaela Gustafson"
date: '2022-07-21'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Create shapefile for prey survey points

In this code I will be importing GPS points for 2019 and 2021 prey surveys and creating a shapefile that is projected into the same CRS as the DEM raster I have.


## LIBRARY

These are the packages used in this code:
```{r library}
library(terra)
library(sf)
library(sp)
library(rgdal)
library(ggplot2)
library(tidyverse)
library(raster)
library(tidyr)
library(dplyr)
library(stringr)
library(rgeos)
library(here)
```

## IMPORT SURVEY LOCATIONS

2019 survey locations are in lat/lon. 2021 survey locations are in UTMs.
I will be converting 2021 points to lat/lon.


### 2019

The recorded lat/lon ended up being in weird locations so I pulled the lat/lon directly from the gps (finalwp_2019.csv) and will be matching the lat/lon waypoints with those recorded in the orignial data sheet ("Point_Count_Env_2019_MG_cleaned.csv")

In 2019 data, there are points not taken for some survey locations so I will be importing the predetermined lat/lon locations for those from the file imported from the GPS (originalpoints_2019.csv)

At the moment, not sure what I used the backup points for...

The script below is taken out of r form:
###########################################################################

points.19 <- read.csv(here("data/Point_Count_Env_2019_MG_cleaned.csv"))
gpx.19 <- read.csv(here("data/finalwp_2019.csv"))
ogpoints.19 <- read.csv(here("data/originalpoints_2019.csv"), skip = 22) # skipping the first 22 rows in the csv because they are extra information imported from the GPS that isn't necessary
#backup.19 <- read.csv(here("data/GPS002_2019backup.csv"), skip = 22)

# There are extra '1' at the end of the waypoint number names, need to remove them so the key id will match when merging dataframes
#backup.19$name = str_remove(backup.19$name, "^0")


# match key id names
gpx.19 <- rename(gpx.19, Waypoint_num = name)
# add leading zero to key id in points.19 to match key id (Waypoint Name)
points.19$Waypoint_num <- str_pad(points.19$Waypoint_num, 3, pad = "0")
# join points and locations from GPS together
points.gpx.join <- left_join(points.19, gpx.19)

# create id key column
points.gpx.join$id <- paste(points.gpx.join$Road_ID, points.gpx.join$UNIT_ID, points.gpx.join$Transect_ID, points.gpx.join$Point_ID, sep = "_")
head(points.gpx.join)


colnames(points.gpx.join)

# only keep columns need
points.2019 <- points.gpx.join[,c(35, 28:29)]
head(points.2019)
# lat lon are empty because I will be pulling that data in from the correct gpx file or original points file. 

# separating out the first 17 surveyed points that don't have matching waypoint
# names in the gps files
missing.points <- points.gpx.join[c(1:17),]

# separate into KOUG and TELL points for naming purposes
miss.points.koug <- missing.points[c(1:11),]
miss.points.tell <- missing.points[c(12:17),]
# Need id names to match those on the points gpx

# Add K or T to column characters respectively so that columns have matching id keys
# and I can bind/join table together 
miss.points.koug$UNIT_ID <- paste("K", miss.points.koug$UNIT_ID, sep = "")
miss.points.tell$UNIT_ID <- paste("T", miss.points.tell$UNIT_ID, sep = "")
miss.points.koug$Transect_ID <- paste("T", miss.points.koug$Transect_ID, sep = "")
miss.points.tell$Transect_ID <- paste("T", miss.points.tell$Transect_ID, sep = "")


# bind all points back together into one dataframe
missing.points <- rbind(miss.points.koug, miss.points.tell)

# create the universal id column in the dataframe of missing points
missing.points$name <- paste(missing.points$UNIT_ID, missing.points$Transect_ID, missing.points$Point_ID, sep = "_")

# join the missing points to dataframe to data taken off GPS to get true lat/lon
# for these points
miss.og.point.join <- left_join(missing.points, ogpoints.19, by = "name")


# only keep columns needed and rename them
colnames(miss.og.point.join)
miss.og.points <- miss.og.point.join[,c(36, 38,39)]
miss.og.points <- rename(miss.og.points, "lat" = "lat.y")
miss.og.points <- rename(miss.og.points, "lon" = "lon.y")
miss.og.points <- rename(miss.og.points, "id" = "name")


### NEED TO CHANGE Column ID NAMES BACK TO MATCH POINTS 2019

koug.names <- as.data.frame(miss.points.koug$id)
koug.names <- rename(koug.names, 'id' = 'miss.points.koug$id')
tell.names <- as.data.frame(miss.points.tell$id)
tell.names <- rename(tell.names, 'id' = 'miss.points.tell$id')
miss.names <- rbind(koug.names, tell.names)


miss.og.points$id <- miss.names$id
str(miss.og.points)

#######################################################################
# remove those incorrect 'missing points' from the dataframe so I can correctly add them in later
points.2019 <- points.2019[complete.cases(points.2019),]

# add corrected points with true lat/lon
points.2019.all <- rbind(miss.og.points, points.2019)

# save CSV of these fixed points and full df for safe keeping
write.csv(points.2019.all, (here("output/points_2019_all_FINAL.csv")))

##############################################################################

## Assign CRS for 2019 data
(pulled from Jenny's version cause I like the way she did it better)

### Now need to import the DEM so we can use the CRS 

```{r dem}
ak_dem <- raster::raster(here("data/dsm_mosaic_AK.tif"))
#plot(ak_dem)
```


```{r 2019crs}

points.2019.all <- read.csv(here("output/points_2019_all_FINAL.csv"))
points.2019.all <- points.2019.all[,-c(1)]

#create an xy dataframe of lat and lon points
xy <- points.2019.all[c(3,2)]
# rename columns 'x' and 'y'
colnames(xy)[1] <- "x"; colnames(xy)[2] <- "y"
# turn into numeric object
xy[,1:2] <- sapply(xy[,1:2], as.numeric)
#turn into coordinates string
coordinates(xy) <- ~ x + y
# now turn into spatial points dataframe matching the newly made saptial points data frame coordinates to the associated data from the original dataframe
spdf <- SpatialPointsDataFrame(coords = xy, data = points.2019.all)
# assing CRS that points were collected in on GPS
proj4string(spdf) <- CRS("+init=epsg:4326")
# reprojecting it to DEM crs first to create extent to crop DEM to more manageable size
points19.proj <- spTransform(spdf, crs(ak_dem)) 
#plot(ak_dem); points(points19.proj)
#writeOGR(points19.proj, dsn=here("shapefiles"), layer="pts19_2" , driver="ESRI Shapefile")

## DON'T THINK I NEED TO DO THIS PART ANYMORE:
# now reproject to crs of land cover data and save
#points19.proj.lc <- spTransform(spdf, crs(r))
#writeOGR(points19.proj.lc, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts19.proj.lc" , driver="ESRI Shapefile")


```


### 2021

Removed R code formatting for this initial cleaning part
############################################################################
##### 2021 #####
points.21 <- read.csv(here("data/Point_Count_Env_2021_MG.csv"))
head(points.21); dim(points.21)

# subset id and utms
colnames(points.21)
points.21 <- points.21[,c(1:4, 19:20)]
points.21$id <- paste(points.21$ROAD_ID, points.21$UNIT_ID, points.21$TRANSECT_ID, points.21$POINT_ID, sep = "_")

head(points.21)
#subset again to keep only final key id and utms
points.21 <- points.21[,c(7, 5:6)]
head(points.21); dim(points.21)
# export for safe keeping
write.csv(points.21, here("output/utms_21_final.csv"))

################################################################################
```{r 21points}
points.21 <- read.csv(here("output/utms_21_final.csv"))
points.21 <- points.21[,-c(1)]


# now do same thing as with 2019 data to assign CRS
xy3 <- points.21[c(2,3)]
colnames(xy3)[1] <- "x"; colnames(xy3)[2] <- "y"
coordinates(xy3) <- ~ x + y
spdf2 <- SpatialPointsDataFrame(coords = xy3, data = points.21)
proj4string(spdf2) <- CRS("+init=epsg:26903")
points21.proj <- spTransform(spdf2, crs(ak_dem))
#plot(ak_dem); points(points21.proj)
#writeOGR(points21.proj, dsn=here("shapefiles"), layer="pts21_2" , driver="ESRI Shapefile")

# AGAIN: Not transforming to the same crs as the land cover data -not sure why I had done this before
#points21.proj.lc <- spTransform(spdf2, crs(r))
#writeOGR(points21.proj.lc, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts21.proj.lc" , driver="ESRI Shapefile")


```

### 2022

```{r 22points}

##### 2022 #####
points.22 <- read.csv(here("data/env22.csv"))
head(points.22); dim(points.22)

# subset id and utms
colnames(points.22)
points.22.sub <- points.22[,c(5, 12:13)]

head(points.22.sub); dim(points.22.sub)
# export for safe keeping
write.csv(points.22.sub, here("output/utms_22_final.csv"))

points.22.sub <- read.csv(here("output/utms_22_final.csv"))
points.22.sub <- points.22.sub[,-c(1)]
# now do same thing as with 2019 data to assign CRS
xy2 <- points.22.sub[c(2,3)]
colnames(xy2)[1] <- "x"; colnames(xy2)[2] <- "y"
coordinates(xy2) <- ~ x + y
spdf2 <- SpatialPointsDataFrame(coords = xy2, data = points.22.sub)
proj4string(spdf2) <- CRS("+init=epsg:26903") # this is the same as the points collected for 2021
points22.proj <- spTransform(spdf2, crs(ak_dem))
#plot(ak_dem); points(points22.proj)
writeOGR(points22.proj, dsn=here("shapefiles"), layer="pts22_3" , driver="ESRI Shapefile")



```


## Now bind all three years together

```{r ptbind}
points19.proj@data$year <- 2019
points21.proj@data$year <- 2021
points22.proj@data$year <- 2022
new <- bind(points19.proj[c(1,4)], points21.proj[c(1,4)], points22.proj[c(1,4)])

# make sure its in dem crs just in case
pts.all <- spTransform(new, crs(ak_dem))

writeOGR(pts.all, dsn=here("shapefiles"), layer="pts_3years_2" , driver="ESRI Shapefile")

plot(ak_dem); plot(pts.all, add = TRUE)

```

# END SCRIPT
