==> 04_SpeciesDataframes.qmd <==
---
title: "Species Dataframes"
format: html
editor: visual
---

# Species Dataframes

In this script I am:

1.  Separating each species into its own observation dataframe
2.  Subsetting ptarmigan species to be male observations only

## Clean workspace

```{r}
# Clean your workspace to reset your R environment
rm( list = ls() )
# Check that you are in the right project folder
getwd()
```

## Library

```{r}
library(dplyr)
library(tidyr)
```

## Load Data

```{r}
all.std <- read.csv("E:/gyrf_analysis/MSgyrfalcon/all_std.csv")
```

# Willow Ptarmigan

## Remove female ptarmigan

```{r}
wipt.male <- data.frame(all.std)

# change all species codes of SQUIR to a 1 and all other species to 0
wipt.male$species[wipt.male$species == "WIPT"] <- 1

wipt.male$species[wipt.male$species != "1"] <- 0

# change count to 0 for non SQUIR
wipt.male$count[wipt.male$species == 0] <- 0

# now to change counts for those obs that were females

unique(wipt.male$sex[wipt.male$species == 1])

# looking at original data notes for M/F sex to know exact number of males and females

wipt.male[3135, 11] = "MMMF"
wipt.male[3137, 11] = "MMF"

wipt.male$sex[wipt.male$sex == "M,F"] <- "MF"
wipt.male$sex[wipt.male$sex == "F,M"] <- "MF"
wipt.male$sex[wipt.male$sex == "M,2U"] <- "MUU"


#remove single F and single U

wipt.male$species[wipt.male$sex == "F"] <- 0
wipt.male$species[wipt.male$sex == "U"] <- 0

wipt.male$count <- as.numeric(wipt.male$count)
wipt.male$sex <- as.factor(wipt.male$sex)

#MF
wipt.male$count[wipt.male$species == 1 & wipt.male$sex == "MF"] <- wipt.male$count[wipt.male$species == 1 &wipt.male$sex == "MF"] - 1

#MMF
wipt.male$count[wipt.male$species == 1 & wipt.male$sex == "MMF"] <- wipt.male$count[wipt.male$species == 1 &wipt.male$sex == "MMF"] - 1

#MMMF
wipt.male$count[wipt.male$species == 1 & wipt.male$sex == "MMMF"] <- wipt.male$count[wipt.male$species == 1 &wipt.male$sex == "MMMF"] - 1

#MUU
wipt.male$count[wipt.male$species == 1 & wipt.male$sex == "MUU"] <- wipt.male$count[wipt.male$species == 1 & wipt.male$sex == "MUU"] - 2

wipt.male$count[wipt.male$species == 0] <- 0

# check for na values
sum(is.na(wipt.male$time_int))
#fix - this was a site where no species were detected
wipt.male[7991, 5] <- 0
# check again
sum(is.na(wipt.male$time_int))

```

## Aggregate time intervals

```{r}
################### 5 TIME INTERVALS #############################
# change time intervals to reflect intervals of 2 minutes
wipt.male2 <- mutate(wipt.male, 
                time_int = ifelse(wipt.male$time_int %in% 0:1, "1", 
                                  ifelse(wipt.male$time_int %in% 2:3, "2", 
                                         ifelse(wipt.male$time_int %in% 4:5, "3",
                                                ifelse(wipt.male$time_int %in% 6:7, "4",
                                                       ifelse(wipt.male$time_int %in% 8:9, "5", "NA"))))))


# look at sample size

wipt.male2$count <- as.integer(wipt.male2$count)
sum(wipt.male2$count) # sample size 485
```

## Get counts and sites for each year

```{r}
# get number of sites with counts each year
wipt.male19 <- subset(wipt.male2, year == "2019" & wipt.male2$count > 0)
sum(wipt.male19$count) # 239 obs in 2019
list(unique(wipt.male19$id)) # 118 sites in 19

wipt.male21 <- subset(wipt.male2, year == "2021" & wipt.male2$count > 0)
sum(wipt.male21$count) # 237 obs in 2021
unique(wipt.male21$id) # 127 sites in 21

wipt.male22 <- subset(wipt.male2, year == "2022" & wipt.male2$count > 0)
sum(wipt.male22$count) # 9 obs in 2022
unique(wipt.male22$id) # 5 sites in 22
###################################################################

```

## Spread dataframe wide

```{r}

# condense counts in same time intervals at each site

str(wipt.male2)
wipt.male2$id <- as.factor(wipt.male2$id)
wipt.male2$time_int <- as.factor(wipt.male2$time_int)
wipt.male2$exact_distance <- as.factor(wipt.male2$exact_distance)
wipt.male2$count <- as.integer(wipt.male2$count)

wipt.male.agg2 <- aggregate(x = wipt.male2$count, 
                      by = list(wipt.male2$id, wipt.male2$time_int),
                      FUN = sum,
                      na.action=NULL)

names(wipt.male.agg2)[names(wipt.male.agg2) == "Group.1"] <- "id"
names(wipt.male.agg2)[names(wipt.male.agg2) == "Group.2"] <- "time_int"
names(wipt.male.agg2)[names(wipt.male.agg2) == "x"] <- "count"

sum(wipt.male.agg2$count) # make sure sample size is still 485



# spread dataframes:

unique(wipt.male.agg2$id) # should end up with 988 levels

wipt.male.wide <- wipt.male.agg2 %>%
  dplyr::select(id, time_int, count) %>%
  spread(key = time_int, value = count, fill = 0)

# double check that sample size is still 485 and sites are 988
sum(wipt.male.wide[,c(2:6)]) 
unique(wipt.male.wide$id)


# checks for missing data

all.std$id <- as.factor(all.std$id)
miss1 <- anti_join(all.std, wipt.male.wide, by = 'id') # clear
miss2 <- anti_join(wipt.male.wide, all.std, by = 'id') # clear

# nothing missing! 



```

## Combine aggregated count data with standardized covariates and other site info

```{r final join}

sub_df <- all.std[,c(2, 12:39)]
df2_unique <- sub_df %>% distinct(id, .keep_all = TRUE)

final_join2 <- left_join(wipt.male.wide, df2_unique, by = 'id')


names(final_join2)[names(final_join2) == "extracted_elev"] <- "elev"
```

## Save dataframe

```{r}
write.csv(final_join2, "E:/gyrf_analysis/MSgyrfalcon/wipt_final_df_28Dec23.csv")

#####################################################

```

# Rock Ptarmigan

## Remove female ptarmigan

```{r}
ropt.male <- data.frame(all.std)

# change all species codes of SQUIR to a 1 and all other species to 0
ropt.male$species[ropt.male$species == "ROPT"] <- 1

ropt.male$species[ropt.male$species != "1"] <- 0

# change count to 0 for non SQUIR
ropt.male$count[ropt.male$species == 0] <- 0

# now to change counts for those obs that were females

unique(ropt.male$sex[ropt.male$species == 1])

# M   M,F   M,U  U  MF

# check how many males are in M,F/U and MF detections

# looking at original data notes for M/F sex to know exact number of males and females

#remove single F and single U

ropt.male$species[ropt.male$sex == "F"] <- 0
ropt.male$species[ropt.male$sex == "U"] <- 0

ropt.male$count <- as.numeric(ropt.male$count)
ropt.male$sex <- as.factor(ropt.male$sex)

# subtract females and unknowns from paired/group observations (all groups were of 2, one male and one female or unknown)

#MF
ropt.male$count[ropt.male$species == 1 & ropt.male$sex == "MF"] <- ropt.male$count[ropt.male$species == 1 &ropt.male$sex == "MF"] - 1

#M,F
ropt.male$count[ropt.male$species == 1 & ropt.male$sex == "M,F"] <- ropt.male$count[ropt.male$species == 1 &ropt.male$sex == "M,F"] - 1

#M,U
ropt.male$count[ropt.male$species == 1 & ropt.male$sex == "M,U"] <- ropt.male$count[ropt.male$species == 1 &ropt.male$sex == "M,U"] - 1


ropt.male$count[ropt.male$species == 0] <- 0
ropt.male$species[ropt.male$count == 0] <- 0


sum(is.na(ropt.male$time_int))
#fix - this was a site where no species were detected
ropt.male[7991, 5] <- 0
# check again
sum(is.na(ropt.male$time_int))
```

## Aggregate time intervals

```{r}
################### 5 TIME INTERVALS #############################
# change time intervals to reflect intervals of 2 minutes
ropt.male2 <- mutate(ropt.male, 
                time_int = ifelse(ropt.male$time_int %in% 0:1, "1", 
                                  ifelse(ropt.male$time_int %in% 2:3, "2", 
                                         ifelse(ropt.male$time_int %in% 4:5, "3",
                                                ifelse(ropt.male$time_int %in% 6:7, "4",
                                                       ifelse(ropt.male$time_int %in% 8:9, "5", "NA"))))))


# look at sample size

ropt.male2$count <- as.integer(ropt.male2$count)
sum(ropt.male2$count) # sample size 55
```

## Get counts and sites for each year

```{r}
# get number of sites with counts each year
ropt.male19 <- subset(ropt.male2, year == "2019" & ropt.male2$count > 0)
sum(ropt.male19$count) # 31 obs in 2019
list(unique(ropt.male19$id)) # 24 sites in 19

ropt.male21 <- subset(ropt.male2, year == "2021" & ropt.male2$count > 0)
sum(ropt.male21$count) # 14 obs in 2021
unique(ropt.male21$id) # 13 sites in 21

ropt.male22 <- subset(ropt.male2, year == "2022" & ropt.male2$count > 0)
sum(ropt.male22$count) # 10 obs in 2022
unique(ropt.male22$id) # 6 sites in 22
###################################################################

```

## Spread dataframe wide

```{r}

# condense counts in same time intervals at each site

str(ropt.male2)
ropt.male2$id <- as.factor(ropt.male2$id)
ropt.male2$time_int <- as.factor(ropt.male2$time_int)
ropt.male2$exact_distance <- as.factor(ropt.male2$exact_distance)
ropt.male2$count <- as.integer(ropt.male2$count)

ropt.male.agg2 <- aggregate(x = ropt.male2$count, 
                      by = list(ropt.male2$id, ropt.male2$time_int),
                      FUN = sum,
                      na.action=NULL)

names(ropt.male.agg2)[names(ropt.male.agg2) == "Group.1"] <- "id"
names(ropt.male.agg2)[names(ropt.male.agg2) == "Group.2"] <- "time_int"
names(ropt.male.agg2)[names(ropt.male.agg2) == "x"] <- "count"

sum(ropt.male.agg2$count) # make sure sample size is still 55



# spread dataframes:

unique(ropt.male.agg2$id) # should end up with 988 levels

ropt.male.wide <- ropt.male.agg2 %>%
  dplyr::select(id, time_int, count) %>%
  spread(key = time_int, value = count, fill = 0)

# double check that sample size is still 55 and sites are 988
sum(ropt.male.wide[,c(2:6)]) 
unique(ropt.male.wide$id)


# checks for missing data

all.std$id <- as.factor(all.std$id)
miss1 <- anti_join(all.std, ropt.male.wide, by = 'id') # clear
miss2 <- anti_join(ropt.male.wide, all.std, by = 'id') # clear

# nothing missing! 
```

## Combine aggregated count data with standardized covariates and other site info

```{r}
sub_df_rock <- all.std[,c(2, 12:39)]
df3_unique <- sub_df_rock %>% distinct(id, .keep_all = TRUE)

final_join_rock <- left_join(ropt.male.wide, df3_unique, by = 'id')


names(final_join_rock)[names(final_join_rock) == "extracted_elev"] <- "elev"
```

## Save dataframe

```{r}
write.csv(final_join_rock, "E:/gyrf_analysis/MSgyrfalcon/ropt_final_df_28Dec23.csv")

#####################################################

```

# Arctic ground squirrel

## Aggregate time intervals

```{r}

squirs <- data.frame(all.std)

# change all species codes of SQUIR to a 1 and all other species to 0
squirs$species[squirs$species == "SQUIR"] <- 1
squirs$species[squirs$species == "AGSQ"] <- 1

squirs$species[squirs$species != "1"] <- 0

# change count to 0 for non SQUIR
squirs$count[squirs$species == 0] <- 0



################### 5 TIME INTERVALS #############################
# change time intervals to reflect intervals of 2 minutes
squirs2 <- mutate(squirs, 
                time_int = ifelse(squirs$time_int %in% 0:1, "1", 
                                  ifelse(squirs$time_int %in% 2:3, "2", 
                                         ifelse(squirs$time_int %in% 4:5, "3",
                                                ifelse(squirs$time_int %in% 6:7, "4",
                                                       ifelse(squirs$time_int %in% 8:9, "5", "NA"))))))


# look at sample size

squirs2$count <- as.integer(squirs2$count)
sum(squirs2$count) # sample size 80

```

## Get counts and sites for each year

```{r}
# get number of sites with counts each year
squirs.19 <- subset(squirs2, year == "2019" & squirs2$count > 0)
sum(squirs.19$count) # 42 obs in 2019
list(unique(squirs.19$id)) # 17 sites in 19

squirs.21 <- subset(squirs2, year == "2021" & squirs2$count > 0)
sum(squirs.21$count) # 35 obs in 2021
unique(squirs.21$id) # 23 sites in 21

squirs.22 <- subset(squirs2, year == "2022" & squirs2$count > 0)
sum(squirs.22$count) # 3 obs in 2022
unique(squirs.22$id) # 2 sites in 22
###################################################################

```

## Spread dataframe wide

```{r}
# condense counts in same time intervals at each site

str(squirs2)
squirs2$id <- as.factor(squirs2$id)
squirs2$time_int <- as.factor(squirs2$time_int)
squirs2$exact_distance <- as.factor(squirs2$exact_distance)
squirs2$count <- as.integer(squirs2$count)

squir.agg2 <- aggregate(x = squirs2$count, 
                      by = list(squirs2$id, squirs2$time_int),
                      FUN = sum,
                      na.action=NULL)

names(squir.agg2)[names(squir.agg2) == "Group.1"] <- "id"
names(squir.agg2)[names(squir.agg2) == "Group.2"] <- "time_int"
names(squir.agg2)[names(squir.agg2) == "x"] <- "count"

sum(squir.agg2$count) # make sure sample size is still 80



# spread dataframes:

unique(squir.agg2$id) # should end up with 988 levels

squir.wide <- squir.agg2 %>%
  dplyr::select(id, time_int, count) %>%
  spread(key = time_int, value = count, fill = 0)

# double check that sample size is still 80 and sites are 988
sum(squir.wide[,c(2:6)]) 
unique(squir.wide$id)


# checks for missing data

all.std$id <- as.factor(all.std$id)
miss1 <- anti_join(all.std, squir.wide, by = 'id') # clear
miss2 <- anti_join(squir.wide, all.std, by = 'id') # clear

# nothing missing! 
```

## Combine aggregated count data with standardized covariates and other site info

```{r}
sub_df_squir <- all.std[,c(2, 12:39)]
df4_unique <- sub_df_squir %>% distinct(id, .keep_all = TRUE)

final_join_squir <- left_join(squir.wide, df4_unique, by = 'id')


names(final_join_squir)[names(final_join_squir) == "extracted_elev"] <- "elev"
```

## Save dataframe

```{r}
write.csv(final_join_squir, "E:/gyrf_analysis/MSgyrfalcon/squir_final_df_28Dec23.csv")

#####################################################

```
